pyproject.toml
[tool.poetry]
name = "insights-puptoo"
version = "0.1.0"
description = "Platform Upload PreProcessor II"
authors = ["Stephen Adams <sadams@redhat.com>"]
packages = [
    { include = "src/puptoo" }
]

[tool.poetry.dependencies]
python = "~3.8"
prometheus-client = "0.7.1"
requests = "2.31.0"
confluent-kafka = "1.5.0"
insights-core = "3.3.18"
app-common-python = "0.2.3"
watchtower = "^1.0.6"
logstash_formatter = "^0.5.17"
minio = "^7.1.11"
urllib3 = "<2"
setuptools = "~65.5.1"

[tool.poetry.dev-dependencies]
pytest = "^5.4.1"
flake8 = "^3.7.9"
freezegun = "^0.3.15"
jsonschema = "^3.2.0"

[tool.poetry.scripts]
puptoo = 'src.puptoo.app:main'
puptoo-run = 'src.puptoo.process.profile:run_profile'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
